Chore(.user): correct incorrectly changed history
